+++ src/App.js
@@ -181,6 +181,15 @@ import SVGAccount from './images/home/account.svg';
 import ProgressTracking, {
   ProgressTrackingDetail,
 } from './containers/ProgressTracking';
+import Profile from './containers/Profile';
+import EditPersonalProfile from './containers/EditProfile';
+import {
+  ListUserChat,
+  ListUserChatNavBar,
+  SearchUserChat,
+  SearchUserChatNavBar,
+  UserChat as AmazingUserChat,
+} from './components/amazingUserChat';
 
 import ListAddressStore from './containers/ListAddressStore';
 /**
@@ -866,6 +875,58 @@ class RootRouter extends Component {
                   </Stack>
                 </Tabs>
 
+                {/* ================ PROFILE ================ */}
+                <Stack key={appConfig.routes.personalProfile}>
+                  <Scene
+                    key={`${appConfig.routes.personalProfile}_1`}
+                    component={Profile}
+                    hideNavBar
+                    back
+                  />
+                </Stack>
+
+                {/* ================ EDIT PROFILE ================ */}
+                <Stack key={appConfig.routes.editPersonalProfile}>
+                  <Scene
+                    key={`${appConfig.routes.editPersonalProfile}_1`}
+                    component={EditPersonalProfile}
+                    title={t('screen.account.editAccountTitle')}
+                    {...navBarConfig}
+                    back
+                  />
+                </Stack>
+
+
+                {/* ================ LIST USER CHAT ================ */}
+                <Stack key={appConfig.routes.listUserChat}>
+                  <Scene
+                    key={`${appConfig.routes.listUserChat}_1`}
+                    title={t('screen.listUserChat.mainTitle')}
+                    component={ListUserChat}
+                    navBar={ListUserChatNavBar}
+                    {...navBarConfig}
+                    back
+                  />
+                  <Scene
+                    key={appConfig.routes.searchUserChat}
+                    component={SearchUserChat}
+                    navBar={SearchUserChatNavBar}
+                    {...navBarConfig}
+                    back
+                  />
+                </Stack>
+
+                {/* ================ AMAZING USER CHAT ================ */}
+                <Stack key={appConfig.routes.amazingUserChat}>
+                  <Scene
+                    key={`${appConfig.routes.amazingUserChat}_1`}
+                    component={AmazingUserChat}
+                    setHeader={this.setHeader.bind(this)}
+                    {...navBarConfig}
+                    back
+                  />
+                </Stack>
+                
                 {/* ================ LIST ADDRESS STORE ================ */}
                 <Stack key={appConfig.routes.listAddressStore}>
                   <Scene