+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- PersonalProfile and UserChat from skv.
+### Fixed
+- Delay setState OrderCart data.
 
 ## [r12.0.4] - 2021-07-09
 ### Added